Pivot : corrige la dernière erreur de compilation (Int vs Int64)
MediaTime.roundedDivision n'a qu'une déclaration dans tout le dépôt,
(Int64, dividedBy: Int64) -> Int64. Array.count est un Int, et Swift
n'applique aucune conversion implicite entre types nominaux distincts, même
sur plateforme 64 bits.

Cinq audits à périmètres disjoints — nouveaux fichiers, modèles de domaine,
câblage, vues, tests — ont convergé sur cette seule erreur. Vérifications
qui ont TENU et n'ont donc rien exigé : l'aplatissement de try? (SE-0230)
dans WaveformStore rend Int64? et non Int64?? ; l'init(from:) en extension
même-fichier préserve le memberwise init et la synthèse de CodingKeys ;
l'appartenance des fichiers neufs à la cible est automatique (dossiers
synchronisés Xcode 16) ; les tuples étiquetés du classificateur
correspondent à ce que lit l'analyseur.

Corrige aussi deux commentaires devenus faux : l'ancien générateur est bien
SUPPRIMÉ et non simplement décâblé — le pivot n'est réversible que par
l'historique —, et la fixture de PaceSelectionStoreTests décrivait encore
trois modes imbriqués alors que l'imbrication §70 est morte avec le pivot.

+++ App/Services/EditScore/PercussiveEditScoreGenerator.swift
@@ -18,9 +18,13 @@
 //  cible de durée, pas de sélection. Ce que le moteur promet devient
 //  vérifiable en une écoute.
 //
-//  L'ancien générateur n'est pas supprimé : il reste compilé et testé, il
-//  n'est simplement plus câblé dans `AppEnvironment`. Le pivot est
-//  réversible.
+//  L'ancien générateur EST SUPPRIMÉ : `DeterministicEditScoreGenerator` et
+//  ses trois fichiers de tests (`EditScoreGeneratorTests`,
+//  `ScoreGenerationPipelineTests`, `ScoreConfigurationWeightsTests`) ne sont
+//  plus dans le dépôt. Le pivot n'est donc PAS réversible par un simple
+//  recâblage d'`AppEnvironment` : il faudrait restaurer ces fichiers depuis
+//  l'historique. `AnchorField` et les poids de `ScoreConfiguration` restent,
+//  eux, en place mais ne sont plus lus par personne.
 //
 //  LA SEULE CONTRAINTE CONSERVÉE
 //  -----------------------------
@@ -143,7 +147,7 @@ struct PercussiveEditScoreGenerator: EditScoreGenerating, Sendable {
             gestures: [],
             averageDuration: durations.isEmpty
                 ? .zero
-                : MediaTime(ticks: MediaTime.roundedDivision(durations.reduce(0, +), dividedBy: durations.count)),
+                : MediaTime(ticks: MediaTime.roundedDivision(durations.reduce(0, +), dividedBy: Int64(durations.count))),
             minimumDuration: MediaTime(ticks: durations.min() ?? 0),
             maximumDuration: MediaTime(ticks: durations.max() ?? 0)
         )
